--- projects/1inch/api.js
@@ -1,5 +1,0 @@
-module.exports = {
-  endpoints: {
-    ethereum: 'https://api.thegraph.com/subgraphs/name/1inch-exchange/oneinch-liquidity-protocol-v2',
-  },
-}
